+++ smerp/src/main/java/com/domino/smerp/salesorder/dto/response/SalesOrderResponse.java
@@ -4,6 +4,7 @@ import com.domino.smerp.salesorder.SalesOrder;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -13,8 +14,8 @@ public class SalesOrderResponse {
     private final Long soId;
     private final Long orderId;
     private final Integer qty;
-    private final Double surtax;
-    private final Double price;
+    private final BigDecimal surtax;
+    private final BigDecimal price;
     private final String remark;
     private final String status;          // ✅ String 유지
     private final LocalDate createdDate;
